Updated Player Movement
Fixed FallingPlatform bug.  Player can now easily jump off of them with
no glitches.

Added a "Platform currentPlatform" variable to the Player class.  This
should hold the currentPlatform the player is colliding with, or null if
the player is not on a platform.

Add a fallingPlatform VerticalState to describe when a player is falling
but on a FallingPlatform.

Smoothed out jumping/falling code.  You can now hold the up arrow and
constantly jump every time you touch the ground/platform.